# docs/swagger.json
{  
    "swagger":"2.0",
    "info":{  
       "description":"This documentation defines API instrumentation for [https://dbullssquad.herokuapp.com/].",
       "version":"1.0.0",
       "title":"E-store",
       "termsOfService":"http://swagger.io/terms/",
       "contact":{  
          "email":"torsami77@gmail.com"
       },
       "license":{  
          "name":"Apache 2.0",
          "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
       }
    },
    "host":"https://dbullssquad.herokuapp.com/",
    "basePath":"api/v1",
...
    "schemes":[  
       "https",
       "http"
    ],
    "paths":{
...
      "/users/updatepassword":{  
         "put":{  
            "tags":[  
               "user"
            ],
            "summary":"updates the user's password",
            "description":"verifies user and update the user's password",
            "operationId":"updatepassword",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {
                  "in":"header",
                  "name":"authorization",
                  "description": "bearer xxxxxxxxxx",
                  "required": true,
                  "schema": {
                     "$ref":"#/definitions/token"
                  }
               },
               {  
                  "in":"body",
                  "name":"body",
                  "description":"update password payload",
                  "required":true,
                  "schema":{  
                        "$ref":"#/definitions/updatepassword"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
...
      "/users/forgotpassword":{
         "post":{  
            "tags":[  
               "user"
            ],
            "summary":"password recovery",
            "description":"",
            "operationId":"forgotpassword",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "name":"email",
                  "in":"body",
                  "description":"The email of user who's password was forgotten",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/forgotpassword"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "type":"string"
                  },
                  "headers":{  
                     "X-Rate-Limit":{  
                        "type":"integer",
                        "format":"int32",
                        "description":"calls per hour allowed by the user"
                     },
                     "X-Expires-After":{  
                        "type":"string",
                        "format":"date-time",
                        "description":"date in UTC when token expires"
                     }
                  }
               },
               "400":{  
                  "description":"Invalid username/password supplied"
               }
            }
         }
      },  
      "/product/add":{  
         "post":{  
            "tags":[  
               "product"
            ],
            "summary":"Adds a new products to the store",
            "description":"",
            "operationId":"addProduct",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"body",
                  "description":"Product payload that needs to be added to the store",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/addProduct"
                  }
               }
            ],
            "responses":{  
               "405":{  
                  "description":"Invalid input"
               }
            }
         }
      },
...
    },
    "securityDefinitions":{  
       "petstore_auth":{  
          "type":"oauth2",
          "authorizationUrl":"https://petstore.swagger.io/oauth/authorize",
          "flow":"implicit",
          "scopes":{  
             "write:pets":"modify pets in your account",
             "read:pets":"read your pets"
          }
       },
       "api_key":{  
          "type":"apiKey",
          "name":"authorization",
          "in":"header"
       }
    },
    "definitions":{
       "token":{
         "type":"apiKey",
         "properties": {
            "token": "bearer xxxxxxxxxx"
         }
       },
...
  	   "updatepassword":{  
          "type":"object",
          "properties":{  
             "password":{  
                "type":"string"
             },
	          "confirmPassword":{  
                "type":"string"
             }
          },
          "xml":{  
             "name":"updatepassword"
          }
       },
...
       "addProduct":{
         "type":"object",
         "properties":{
            "image_url":{
               "type":"string"
            },
            "price": {
               "type":"integer"
            },
            "description":{
               "type":"string"
            }
         },
         "xml":{
            "name":"signin"
         }
       },
       "forgotpassword": {
            "type": "object",
            "properties":{
               "email": {
                  "type":"string"
               }
            }
       },
...
    },
...
 }